Split qasm parser
